requirements.txt
discord.py
sqlalchemy
pymysql
aiomysql
python-dotenv
audioop-lts
